refactor : UserEntity > User로 엔티티명 변경
