+++ src/main/java/com/better/CommuteMate/application/schedule/dtos/WorkScheduleCommand.java
@@ -1,7 +1,7 @@
 package com.better.CommuteMate.application.schedule.dtos;
 
 import com.better.CommuteMate.domain.schedule.entity.WorkSchedule;
-import com.better.CommuteMate.domain.user.entity.UserEntity;
+import com.better.CommuteMate.domain.user.entity.User;
 
 import java.time.LocalDateTime;
 
@@ -10,7 +10,7 @@ public record WorkScheduleCommand(
         LocalDateTime start,
         LocalDateTime end) {
 
-    public static WorkSchedule toEntity(WorkScheduleCommand command, UserEntity user) {
+    public static WorkSchedule toEntity(WorkScheduleCommand command, User user) {
         return WorkSchedule.builder()
                 .user(user)
                 .date(command.start.toLocalDate())
