--- src/main/java/com/better/CommuteMate/domain/auth/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.better.CommuteMate.domain.auth.repository;
-
-import com.better.CommuteMate.domain.user.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, String> {
-    Optional<UserEntity> findByEmail(String email);
-    boolean existsByEmail(String email);
-}
