+++ src/main/java/com/better/CommuteMate/domain/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.better.CommuteMate.domain.user.repository;
+
+import com.better.CommuteMate.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
+}
